Rooms Allow Multiple People
Re-read the instruction and multiple people are allows in rooms, but no
hallways. Refactored room code so it allows multiple people in and
shifts the token drawing depending on how many people are in the room.

Also added skip turn button so the game won't end if all hallways are
blocked.